Provide settings for button layout
The DecorationSettings have two new properties for left and right
decoration buttons. In addition the DecorationButtonGroup gains a
new ctor which can specify which group it is (left or right) and
accepts a std::function to create a DecorationButton. Thus the
Decoration does not have to care about creating the buttons and
adding it to the correct layout. The DecorationButtonGroup takes
care of it.