DecorationBridge is no longer a singleton
The DecorationBridge needs to be passed to the DecoratedClient and
Deocration ctor, so that they don't need to call the singleton
any more. The challenge is that we are restricted in what can be
passed to Decoration's ctor if we want to create a Decoration with
the help of KPluginFactory. As a solution the DecorationBridge can
be passed in through the QVariantList which is passed to a created
instance by KPluginFactory. This means for inheriting classes that
they need to pass on the QVaraintList argument.

For backends to implement the DecorationBridge it is now important
to pass their DecorationBridge in the QVariantList passed as args
to KPluginFactory. It's expected to be in a QVariantMap under key
"bridge".