Decoration holds a QSharedPointer of DecoratedClient
Decoration::client returns a QWeakPointer. It can be guaranteed that
the pointer is valid as long as the Decoration is alive. That is in
a subclass one can directly access the pointer. If the client is
added to e.g. a hash it's suggested to promote it to QSharedPointer
when used.