Restore ABI of libkdecorations2.so
Restore ABI of libkdecorations2.so

requestShowWindow() was removed. Restore with an implementation that
sends an invalid QRect so that we don't need two versions throughout the
private API as well. A little bit hacky, but it should suffice
well-enough for deprecated code.

This should be ABI compatiable with the existing released
libkdecorations2.so.5