Introduce Decoration::damaged signal
This signal is needed to support various server decoration renderers. In
kwin, we'd like to get rid of the DecorationRenderer in the composited
code path in favor of a unified texture atlas. However, on X11, we still
need to support the code path where compositing is off.

This change introduces the Decoration::damaged() signal to make server
side decoration rendering machinery in the compositor more flexible. The
new signal is similar to the SurfaceInterface::damaged() signal.