Deprecate DecoratedClient::windowId() and DecoratedClient::decorationId()
They are very X11-specific and can't be made work properly on Wayland.
Breeze uses windowId() to embed size grip into the client window and get
the info about the window. Size grip is not something critical
and it's disabled by default, so I think that we can drop it. Providing
X11 WId to get window info is not a valid usecase.