Expose device pixel ratio to clients
Expose display pixel ratio to KDecoration so it can be used by
KDecoration and any themes that uses it, to modify border sizes and
other variables based on the dpr. This is useful for current solutions
to fix the gaps between borders bug, without making that fix affect
unscaled windows, where it causes different issues.